refactor: 알맞는 레이어로 파일 이동
