+++ src/pages/post-manager/ui/PostsManagerPage.tsx
@@ -1,15 +1,15 @@
-import { usePosts } from "@entities/comment/model/usePosts.ts"
-import AddCommentDialog from "@pages/post-manager/ui/dialogs/AddCommentDialog.tsx"
-import AddPostDialog from "@pages/post-manager/ui/dialogs/AddPostDialog.tsx"
-import EditCommentDialog from "@pages/post-manager/ui/dialogs/EditCommentDialog.tsx"
-import EditPostDialog from "@pages/post-manager/ui/dialogs/EditPostDialog.tsx"
-import PostDialog from "@pages/post-manager/ui/dialogs/PostDialog.tsx"
-import UserDialog from "@pages/post-manager/ui/dialogs/UserDialog.tsx"
-import FilterBar from "@pages/post-manager/ui/FilterBar.tsx"
-import PostPagination from "@pages/post-manager/ui/PostPagination.tsx"
+import AddCommentDialog from "@features/comment-update/ui/AddCommentDialog.tsx"
+import EditCommentDialog from "@features/comment-update/ui/EditCommentDialog.tsx"
+import FilterBar from "@features/post-filter/ui/FilterBar.tsx"
+import PostPagination from "@features/post-pagination/ui/PostPagination.tsx"
+import AddPostDialog from "@features/post-update/ui/dialogs/AddPostDialog.tsx"
+import EditPostDialog from "@features/post-update/ui/dialogs/EditPostDialog.tsx"
 import Posts from "@pages/post-manager/ui/Posts.tsx"
+import { usePosts } from "@shared/hooks/usePosts.ts"
 import { useLayoutStore } from "@shared/model/store.ts"
 import { Button, Card, CardContent, CardHeader, CardTitle } from "@shared/ui"
+import PostDialog from "@widgets/ui/dialogs/PostDialog.tsx"
+import UserDialog from "@widgets/ui/dialogs/UserDialog.tsx"
 import { Plus } from "lucide-react"
 import { useCallback, useEffect } from "react"
 import { useLocation, useNavigate } from "react-router-dom"
