+++ src/widgets/ui/dialogs/PostDialog.tsx
